+++ src/Service/WithdrawServiceInterface.php
@@ -8,11 +8,11 @@ use Jorijn\Bitcoin\Dca\Model\CompletedWithdraw;
 
 interface WithdrawServiceInterface
 {
-    public function withdraw(string $asset, float $amountToWithdraw, string $addressToWithdrawTo): CompletedWithdraw;
+    public function withdraw(string $asset, int $amountToWithdraw, string $addressToWithdrawTo): CompletedWithdraw;
 
-    public function getAvailableBalance(string $assetToWithdraw): float;
+    public function getAvailableBalance(string $assetToWithdraw): int;
 
-    public function getWithdrawFee(string $asset, float $amountToWithdraw, string $addressToWithdrawTo): float;
+    public function getWithdrawFee(string $asset, int $amountToWithdraw, string $addressToWithdrawTo): int;
 
     public function getWithdrawFeeInSatoshis(): int;
